Reorganizing layers components
