+++ package.json
@@ -32,7 +32,7 @@
   },
   "devDependencies": {
     "babel-core": "^6.0.0",
-    "babel-eslint": "^7.2.2",
+    "babel-eslint": "^7.2.3",
     "babel-loader": "^6.0.0",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-preset-latest": "^6.0.0",
