+++ src/components/layer/CanvasLayer.vue
@@ -1,8 +1,12 @@
-<template></template>
+<template>
+  <div style="display: none">
+    <slot></slot>
+  </div>
+</template>
 
 <script>
-  import bus from '../messageBus';
-  import layerEvents from '../lib/layerEvents';
+  import bus from '../../messageBus';
+  import layerEvents from '../../lib/layerEvents';
 
 
   export default {
@@ -14,9 +18,9 @@
         type: Array,
         required: true
       },
-      url: {
-        type: String,
-        required: true
+      animate: {
+        type: Boolean,
+        default: true
       },
 
       // mapbox layer style properties
@@ -31,7 +35,7 @@
       maxzoom: Number,
       // filter: Object,
       // layout: Object,
-      paint: Object,
+      // paint: Object,
 
       // mapbox layer options
       before: Object,
@@ -67,10 +71,15 @@
     },
 
     mounted() {
+      if (this.$slots.default[0].tag !== 'canvas') {
+        throw new Error(`Error in map layer component with source id "${this.sourceId}" and layer id "${this.layerId}"
+          You need to add canvas element as child of canvas layer.`)
+      }
       let source = {
-        type: 'image',
-        url: this.url,
-        coordinates: this.coordinates
+        type: 'canvas',
+        coordinates: this.coordinates,
+        animate: this.animate,
+        canvas: this.$slots.default[0].data.attrs.id
       }
       // We wait for "load" event from map component to ensure mapbox is loaded and map created
       bus.$on('mgl-load', map => {
@@ -117,6 +126,9 @@
       },
       mapLayer() {
         return this.map.getLayer(this.layerId);
+      },
+      canvas() {
+        return this.map.getSource(this.sourceId).getCanvas();
       }
     },
 
@@ -129,13 +141,13 @@
         if (this.initial) return;
         this.map.setLayerZoomRange(this.layerId, this.minzoom, val)
       },
-      paint(val) {
-        // FIXME: save initial state and replace only changed fields?
-        if (this.initial) return;
-        val.keys().forEach(key => {
-          this.map.setPaintProperty(this.layerId, key, val);
-        });
-      },
+      // paint(val) {
+      //   // FIXME: save initial state and replace only changed fields?
+      //   if (this.initial) return;
+      //   val.keys().forEach(key => {
+      //     this.map.setPaintProperty(this.layerId, key, val);
+      //   });
+      // },
       coordinates(val) {
         if (this.initial) return;
         this.map.setCoordinates(val);
@@ -213,8 +225,8 @@
         layer.metadata = this.metadata
 
         this.map.addLayer(layer, this.before);
-        this.$emit('mgl-layer-added', this.layerId);
-        bus.$emit('mgl-layer-added', this.layerId);
+        this.$emit('mgl-layer-added', { component: this, layerId: this.layerId });
+        bus.$emit('mgl-layer-added', { component: this, layerId: this.layerId });
       },
 
       move(beforeId) {
